backend/requirements.txt
# 国内股票行情网关 - Python 依赖
# Python 3.10+

# FastAPI 框架
fastapi==0.109.2
uvicorn[standard]==0.27.1

# 异步 HTTP 客户端
httpx==0.26.0
aiohttp==3.9.3
httpx==0.26.0

# 异步 TDengine 客户端 (taos-ws)
taospy==2.8.8

# 数据验证
pydantic==2.6.1
pydantic-settings==2.1.0

# 配置文件
PyYAML==6.0.1

# 轻量级 SQLite (如需本地配置持久化)
aiosqlite==0.19.0

# 异步任务调度
python-multipart==0.0.9

# 日期时间处理
python-dateutil==2.8.2

# 日志
loguru==0.7.2

# 缓存
cachetools==5.3.2

# 预加载 typing_extensions (加速)
typing-extensions==4.9.0
